WIP: moving commands to objects
