+++ test/fixtures/npm-ls.txt
@@ -1,0 +1,11 @@
+sustain@1.1.0 /Users/joshua/projects/sustain
+├── bitcoin-regex@1.1.0
+├── chai@2.3.0
+├── commist@1.0.0
+├── mocha@2.2.4
+├── sinon@1.14.1
+├── sinon-chai@2.7.0
+├── sorted-object@1.0.0
+├── standard@3.3.0
+└── xtend@4.0.0
+
