+++ test/init.js
@@ -1,12 +1,12 @@
 /* global describe, it, afterEach, beforeEach */
 var expect = require('chai').expect
 
-var CommandTester = require('./lib/shared')
+var CommandTester = require('./lib/command-tester')
 var validAddress = require('./fixtures/test-address')
 var packageFixture = require('./lib/package-fixture-manager')
 
-var init = require('../commands/init')
-var sharedTester = new CommandTester(init, [validAddress])
+var InitCommand = require('../commands/init')
+var sharedTester = new CommandTester(new InitCommand(), [validAddress])
 
 process.chdir(__dirname)
 
@@ -18,14 +18,14 @@ describe('init', function () {
   sharedTester.handlesInvalidArgs([])
 
   it('errors when bitcoin address is invalid', function (done) {
-    init('asdf', function (err) {
+    new InitCommand().call('asdf', function (err) {
       expect(err.message).to.equal('Given address is invalid.')
       done()
     })
   })
 
   it('updates package.json with valid address', function (done) {
-    init(validAddress, function () {
+    new InitCommand().call(validAddress, function () {
       expect(packageFixture.read().sustain).to.deep.equal({
         address: validAddress // random test address
       })
