+++ test/update.js
@@ -1,9 +1,11 @@
-/* global describe, it, afterEach */
+/* global describe, afterEach, it */
 var expect = require('chai').expect
 
+var CommandTester = require('./lib/command-tester')
 var packageFixture = require('./lib/package-fixture-manager')
 
-var update = require('../commands/update')
+var UpdateCommand = require('../commands/update')
+var sharedTester = new CommandTester(new UpdateCommand(), [])
 
 process.chdir(__dirname)
 
@@ -11,13 +13,12 @@ describe('update', function () {
   afterEach(function () {
     packageFixture.cleanup()
   })
-  it('displays error if no sustain field in package.json', function (done) {
-    packageFixture.setup('empty')
-    update(function (err) {
-      expect(err.message).to.equal('No sustain data in package.json.')
-      done()
-    })
+  sharedTester.requiresPackageFile()
+  sharedTester.requireSustainField()
 
+  it('loads installed packages', function () {
+    packageFixture.setup('basic')
+    expect(true).to.equal(true)
   })
 
 })
